updated Web Api

+++ BackendCode/Backend_API/Backend_API/Models/DTO/PlayerDTO.cs
@@ -2,6 +2,6 @@
 {
     public class PlayerDTO
     {
-        public long RoomId { get; set; }
+        public int RoomId { get; set; }
     }
 }
